output/V-KYC 1st successful/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
# Optional: For Prometheus metrics
prometheus-fastapi-instrumentator==6.2.0
# For testing
pytest==8.2.1
httpx==0.27.0
locust==2.21.0
# For code quality
flake8==7.0.0
black==24.4.2
isort==5.13.2